Filter table based on the fuzzy search
